fix: useFixedPriceSalePurchaseByBuyer hook
