+++ src/hooks/useBids.ts
@@ -1,6 +1,5 @@
 // External
 import { useDispatch, useSelector } from 'react-redux'
-import { BigNumber } from 'ethers'
 import dayjs from 'dayjs'
 import { useEffect } from 'react'
 import { useWeb3React } from '@web3-react/core'
@@ -24,19 +23,6 @@ interface UseBidsReturn {
   error: Error | null
   bids: SaleBid[]
   totalBids: SaleBid[]
-}
-
-/** sums up all the purchases of a single user from a single sale and returns the status, total amount and account string*/
-export const aggregatePurchases = (bids: SaleBid[], account: string | null | undefined) => {
-  const reduceTotalAmount = bids.reduce((accumulator: BigNumber, purchases: any) => {
-    return BigNumber.from(accumulator).add(purchases.amount)
-  }, BigNumber.from(0))
-
-  return {
-    buyer: account!,
-    amount: reduceTotalAmount,
-    status: bids.length > 0 ? bids[0].status : undefined,
-  }
 }
 
 export function useBids(saleId: string, saleType: SaleType): UseBidsReturn {
